requirements.txt
fastapi==0.108.0
uvicorn==0.26.0
SQLAlchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4